api_amazon/requirements.txt
django
djangorestframework
djangorestframework-jwt
psycopg2-binary
requests
drf_yasg
Pillow
